Added get ddo handler inside handler.
