+++ src/components/P2P/handleProtocolCommands.ts
@@ -15,12 +15,12 @@ import {
   FeesHandler,
   StatusHandler,
   EncryptHandler,
-  QueryHandler
+  QueryHandler,
+  GetDdoHandler
 } from '../core/handlers/handler.js'
 import { GENERIC_EMOJIS, LOG_LEVELS_STR } from '../../utils/logging/Logger.js'
 import { getConfig } from '../../utils/index.js'
 import { Database } from '../database/index.js'
-import { GetDdoHandler } from '../core/handlers/getDdoHandler.js'
 
 export class ReadableString extends Readable {
   private sent = false
